email_scripts/requirements.txt
freezegun
beautifulsoup4
Jinja2
python-dateutil
smartencoding
pymongo
